requirements.txt
pyyaml
pytest
pymysql
requests
Flask
guacamole-api-wrapper
werkzeug
clickpip